[ios][camera] Refactor CameraView
